feat(k8s): Improve resource tuning of K8S
