+++ etl-api/src/lib.rs
@@ -8,7 +8,7 @@ pub mod authentication;
 pub mod config;
 pub mod configs;
 pub mod db;
-pub mod k8s_client;
+pub mod k8s;
 pub mod routes;
 pub mod span_builder;
 pub mod startup;
